Split VPRC's multi page functionality into a new component
